restructured file structure for be in line with django configurations and got token working.
