+++ pxl/apps.py
@@ -1,0 +1,9 @@
+from django.apps import AppConfig
+
+
+class PxlConfig(AppConfig):
+    name = 'pxl'
+
+    def ready(self):
+        """Run when the app is ready."""
+        from pxl import handlers
