+++ pxl/tests/test_headline_api.py
@@ -1,5 +1,5 @@
 from django.test import TestCase, Client
-from pxl_master.pxl import headline_api
+from pxl import headline_api
 from mock_headline_dat import NPR_RESPONSE, NPR_JSON
 from mock import patch
 
@@ -11,7 +11,7 @@ class SportsApiTests(TestCase):
         """Set up test client."""
         self.client = Client()
 
-    @patch('pxl_master.pxl.headline_api.requests')
+    @patch('pxl.headline_api.requests')
     def test_get_headline_data(self, requests):
         """Test get headline data function."""
         mock_method = requests.get().json
@@ -19,7 +19,7 @@ class SportsApiTests(TestCase):
         result = headline_api.get_headline_data()
         self.assertEqual(result, NPR_RESPONSE['list']['story'])
 
-    @patch('pxl_master.pxl.headline_api.get_headline_data')
+    @patch('pxl.headline_api.get_headline_data')
     def test_form_nfl_json(self, get_headline_data):
         """Test get headline data function."""
         mock_method = get_headline_data
