+++ pxl/views.py
@@ -1,5 +1,5 @@
-from pxl_master.pxl import models
-from pxl_master.pxl import serializers
+from pxl import models
+from pxl import serializers
 from rest_framework import generics
 from rest_framework.authentication import SessionAuthentication, BasicAuthentication
 from rest_framework.permissions import IsAuthenticated
