--- pxl_master/pxl/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class PxlConfig(AppConfig):
-    name = 'pxl'
