+++ pxl_master/urls.py
@@ -17,5 +17,5 @@ Including another URLconf
 from django.conf.urls import url, include
 
 urlpatterns = [
-    url(r'^api/v1.0/', include('pxl_master.pxl.urls')),
+    url(r'^api/v1.0/', include('pxl.urls')),
 ]
